UPDATE: implement InputTextMultiline callback for replacing tab with spaces
Although we can retrieve callback data passed from the widget
`InputTextMultiline` by wrapping `ImGuiInputTextCallbackData`
using `ctypes`. However, we cannot manipulate content in widget
since those callback data are not the actual data used in that
widget.

Because the data we get are actually Python objects *translated*
by the ctypes struct `InputTextCallbackDataStruct`, their memory
addresses are not the same as those data created in C++ (ImGUI).

Hence that it seems not possible to manipulate widget data by
simply passing a callback method made in Python currently. Though
there is still an approach worthy to try: wrap those functions
defined in `ImGuiInputTextCallbackData` struct (e.g. `DeleteChars`,
`InsertChars`...) by `ctypes` to make user able to call them
through a callback made in Python. We've tried it, but no luck
so far.

Therefore, we re-implement callback functions in Cython in order
to interact with those callback data. Though this approach is
less flexiable but it works well.