MAINT: SelectionStart, SelectionEnd and CursorPos were not set correctly after indenting/unindenting paragraph
Considering our implementaion of paragraph indentation and
unindentation, we will split the text into 3 parts before
inserting or removing indentation:
- `head`: Buf[:SelectionStart]
- `text`: Buf[SelectionStart:SelectionEnd]
- `tail`: Buf[SelectionEnd:]

Since length of `text` will always be changed after being indented
or unindented, we should keep `SelectionStart`, `SelectionEnd`
and `CursorPos` enclosing the selected paragraph. Otherwise,
range of selected paragraph will be detected incorrectly and make
text be splitted incorrectly (i.e. `head`, `text`, `tail`). And it
eventually lead to an incorrect indentation/unindentation.

And since we have already dealt with the problem of reversed
selection, things become easiler to understand that:
- `SelectionStart` is still the same value of `len(head)`
- `SelectionEnd` should be `len(head)` + `len(text)`
These 2 values should be swapped if selection is reversed.

As for `CursorPos`, it's commonly set to the location of last
character of selected paragraph in text editors. Therefore, it
should be the same value as `len(head)` + `len(text)`.

Besides, there there is a problem while we are going to insert
indentation characters: the Last line of paragrah won't be indented
if there is no new-line character in the last line. To solve this,
the comparison of `end` position and length of `data.Buf` should be
`<=` instead of `<`.