MAINT: change key combination for reverse indentation from "ctrl+tab" back to "shift+tab"
In commit 7c6aaff, key combination for reverse indentation (it was
called `unindent` in code) was set to "ctrl+tab" because KEY_TAB
would not be detected when KEY_SHIFT is pressed. It was considered
as an issue comes from `imgui`.

However, since we found there was an issue about key stroke handling
in the integrated module `pyglet` of `pyimgui`, we came back to this
and found that KEY_TAB can actually be detected even when KEY_SHIFT
is pressed.

Therefore, key combination for reverse indentation is now changed
to "shift+tab" since the latter one is more commonly used.

See also the following link for the patch:
https://github.com/NaleRaphael/codememo/commit/7e62fbb0133c4237e5f5a0ed0174643416bc42d6